requirements.txt

django
request